fix bugs in delete course

+++ CRS-JEDI-JAVA-Development-POS/src/com/flipkart/dao/AdminDAOImpl.java
@@ -58,9 +58,18 @@ public class AdminDAOImpl implements AdminDAO {
         Connection conn = null;
         PreparedStatement stmt = null;
 
+
+
         try{
             Class.forName(Dao.JDBC_DRIVER);
             conn = DriverManager.getConnection(Dao.DB_URL,Dao.USER,Dao.PASS);
+
+            String sql1 = "Select * from courses where courseId = ?";
+            stmt = conn.prepareStatement(sql1);
+            stmt.setInt(1, courseID);
+            ResultSet rs = stmt.executeQuery();
+
+            if(rs.next()==false)  {throw new CourseNotFoundException(courseID);}
 
             String sql = "DELETE FROM Catalog WHERE CourseId = ?";
             stmt = conn.prepareStatement(sql);
@@ -77,16 +86,18 @@ public class AdminDAOImpl implements AdminDAO {
             sql = "DELETE FROM Courses WHERE CourseID = ?";
             stmt = conn.prepareStatement(sql);
             stmt.setInt(1, courseID);
-            int row = stmt.executeUpdate();
-            if (row == 0) {
-                System.out.println(courseID + " not in catalog!");
-                throw new CourseNotFoundException(courseID);
-            }
+            stmt.executeUpdate();
+
+            System.out.println("Course removed successfully.");
 
         } catch(SQLException se){
             //Handle errors for JDBC
             throw new CourseNotDeletedException(courseID);
-        }catch(Exception e){
+        } catch (CourseNotFoundException e)
+        {
+            System.out.println(e.getMessage());
+            return;
+        } catch(Exception e){
             //Handle errors for Class.forName
             e.printStackTrace();
         }finally{
